Allow running Rediseen in daemon mode (#7)
Allow users to apply flag "-d" to run Rediseen in background, and provide command "stop" to stop Rediseen running in background.

Implemented by storing the PID of the Rediseen process in a file (using os.TempDir() for default PID file location, as it's safer and can ensure feature works on all operating platforms). Users are also allowed to specify "-pidfile" if they want to have a different location for the PID file, or they want to run multiple Rediseen instances.